Add files via upload
This codebase contains a couple of files that create a complete User module from the list to detail. You are free to use these files and make all the changes you need in order to make is used for your work. Please also raise a pull request if you wish to make improvements.